Fixed various problems with the drawing and interactivity of time line at high zoom levels caused by insufficient floating point precision.

+++ src/vogleditor/vogleditor_apicalltimelinemodel.h
@@ -36,7 +36,7 @@ class vogleditor_apiCallTimelineModel : public vogleditor_timelineModel
 public:
     explicit vogleditor_apiCallTimelineModel(vogleditor_apiCallTreeItem *pRootApiCall);
     void refresh();
-    float absoluteToRelativeTime(uint64_t time);
+    double absoluteToRelativeTime(uint64_t time);
 
 private:
     unsigned int randomRGB();
